requirements.txt
# Core Airflow
apache-airflow==2.8.0
apache-airflow-providers-snowflake==5.1.1
apache-airflow-providers-common-sql==1.10.0

# Data transformation
dbt-core==1.6.6
dbt-snowflake==1.7.1

# Data processing
pandas==2.1.4
requests==2.31.0

# Snowflake connectivity
snowflake-connector-python==3.6.0
snowflake-sqlalchemy==1.5.1

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
cryptography==41.0.7
pyarrow==14.0.1
numpy==1.24.3 